Move some classes to their correct module.
