+++ features/rageshake/impl/src/main/kotlin/io/element/android/features/rageshake/impl/crash/CrashDataStore.kt
@@ -5,7 +5,7 @@
  * Please see LICENSE files in the repository root for full details.
  */
 
-package io.element.android.features.rageshake.api.crash
+package io.element.android.features.rageshake.impl.crash
 
 import kotlinx.coroutines.flow.Flow
