+++ features/rageshake/impl/src/main/kotlin/io/element/android/features/rageshake/impl/rageshake/RageShake.kt
@@ -5,7 +5,7 @@
  * Please see LICENSE files in the repository root for full details.
  */
 
-package io.element.android.features.rageshake.api.rageshake
+package io.element.android.features.rageshake.impl.rageshake
 
 interface RageShake {
     /**
